# Package collections: improve search performance