--- a/Sources/PackageCollections/CMakeLists.txt
+++ b/Sources/PackageCollections/CMakeLists.txt
@@ -23,6 +23,7 @@ add_library(PackageCollections
   Storage/PackageCollectionsSourcesStorage.swift
   Storage/PackageCollectionsStorage.swift
   Storage/SQLitePackageCollectionsStorage.swift
+  Storage/Trie.swift
   API.swift
   PackageCollections.swift
   PackageCollections+Configuration.swift